pyproject.toml
[tool.poetry]
name = "electro-store"
version = "0.1.0"
description = ""
authors = ["KonstantinMikheev <mikhey17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.8"
ipython = "^8.31.0"
pillow = "^11.1.0"
django-stubs = "^5.1.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
